Break runtime errors example into sections
